requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
ffmpeg-python==0.2.0
future==0.18.2
idna==3.4
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0
pyTelegramBotAPI==4.7.1
requests==2.28.1
urllib3==1.26.12
